+++ sdk/src/protocols/schema.ts
@@ -13,6 +13,7 @@ import * as rocketPool from "./rocketPool/schema"
 import * as spark from "./spark/schema"
 import * as stader from "./stader/schema"
 import * as stakeWiseV2 from "./stakeWise/v2/schema"
+import * as stakeWiseV3 from "./stakeWise/v3/schema"
 import * as uniswapV3 from "./uniswap/v3/schema"
 import { ProtocolSchemas } from "../types"
 
@@ -34,6 +35,7 @@ export const eth = {
   spark: spark.eth,
   stader: stader.eth,
   stakewise_v2: stakeWiseV2.eth,
+  stakewise_v3: stakeWiseV3.eth,
   uniswap_v3: uniswapV3.eth,
 } satisfies Record<string, ProtocolSchemas>
 
@@ -43,6 +45,7 @@ export const gno = {
   balancer: balancer.gno,
   cowswap: cowSwap.gno,
   spark: spark.gno,
+  stakewise_v3: stakeWiseV3.gno,
 } satisfies Record<string, ProtocolSchemas>
 
 export const arb1 = {