+++ sdk/src/protocols/stakeWise/v3/_ethVaults.ts
@@ -1,0 +1,216 @@
+// This file is auto-generated. Do not edit!
+
+export default [
+  {
+    id: "0xac0f906e433d58fa868f936e8a43230473652885",
+    name: "Genesis",
+    description:
+      "This is the vault for the StakeWise V2 pool. After migrating from StakeWise V2 to V3, you will be staking in this vault. The node operators in this vault are: Cryptomanufaktur, Deutsche Telekom, Finoa, StakeWise Labs.",
+  },
+  {
+    id: "0xe6d8d8ac54461b1c5ed15740eee322043f696c08",
+    name: "Chorus One - MEV Max",
+    description:
+      "Chorus One’s ground-breaking MEV research ensures the highest yields with top-tier security and enterprise-level infrastructure. Start staking ETH today.",
+  },
+  {
+    id: "0xa8cb2e35f594b308726b242136f0a8995bf19e97",
+    name: "Easy Stake - Powered by SSV",
+    description:
+      "3 out of 4 Fault Tolerance  - SSV Operators 491, 770, 1002, 1003 - Solo node runner since April 2023 -  Minority clients Nethermind/Nimbus, Besu/Teku, Erigon/Lodestar, Nethermind/Prysm - MEV enabled - High-end NUC with UPS",
+  },
+  {
+    id: "0x8a93a876912c9f03f88bc9114847cf5b63c89f56",
+    name: "Stakingverse - Ethereum",
+    description:
+      "Welcome to The Stakingverse!\n\nA premier staking provider with competitive fees, shared MEV and state-of-the-art infrastructure provided by our partners at Digital Ocean. \n\nAs an official StakeWise partner, we are dedicated to perfecting the secure and decentralized nature of the V3 Protocol to multiple chains. And with a team of highly-experienced professionals, who have been on the front-line of blockchain technology for many years now, you will get the best service, APY and security in liquid staking, all while being non-custodial.\n\nStake your LYX, ETH and GNO with The Best - stakingverse.io",
+  },
+  {
+    id: "0xb36fc5e542cb4fc562a624912f55da2758998113",
+    name: "Serenita - Powered by SSV",
+    description:
+      "Serenita provides enterprise-grade staking services, prioritizing optimal performance while mitigating the risks associated with using majority clients. Our exclusive focus is on the Ethereum ecosystem and we are dedicated to its long-term sustainability and neutrality.\n\nThe validators in this Vault are powered by ssv.network's DVT technology, which ensures resiliency against client bugs and reduces validator downtime. Additionally, validators are eligible for SSV's Incentivized Mainnet Program, boosting rewards by up to 30%! The program rewards will be retroactively distributed among Vault depositors at the program's conclusion at the end of 2024.\n\nFor a deeper understanding of our operational approach, visit our website at serenita.io",
+  },
+  {
+    id: "0x8216e50b1dd76faeb1ff4cafaba6790ea71910df",
+    name: "Data Nexus",
+    description:
+      "Independent Ethereum validators with owned hardware, MEV boost, and minority clients for robust decentralization. Empowering the Ethereum network with integrity.",
+  },
+  {
+    id: "0x663580b3edad914d0b59cea88616f06278d42bb2",
+    name: "𓃵 onchainaustria.at",
+    description:
+      "onchainaustria is a registered Austrian association and thriving crypto community. We operate as a node operator with a low fee structure and a highly redundant technical setup.\n\nWe are offering native crypto staking. Our members benefit from a collaborative environment where we share knowledge and optimize crypto investments.\n\nhttps://onchainaustria.at/",
+  },
+  {
+    id: "0xb46dba422bcf7f113da2eeb41c0e2cc6298d6bba",
+    name: "Blockscape Super Early Bird",
+    description:
+      "Institutional-grade staking for everyone made by Blockscape in Switzerland. As an early staker, you benefit from a lower commission fee of just 4%, while still enjoying the same levels of security and reliability. We are pioneering the distributed validator technology (DVT) in Stakewise with SSV.\nOur unique setup provides highest standards in security and reliability through a distributed DVT setup located in Canada, Germany, Singapore, and France. With more than 10.000 stakers and over $500.000.000 under management, Blockscape is one of the most reliable staking providers in the world.\nMoreover, we offer dedicated vaults for institutions with specific requirements. Reach out to us at: partner@blockscape.network\nor visit https://blockscape.network for more information.",
+  },
+  {
+    id: "0x5610297eb32542b8f13378fe7d783dcb1ac3b1a1",
+    name: "SenseiNode",
+    description:
+      "SenseiNode is the first institutional-grade blockchain infrastructure service provider in Latin America. We automate and simplify the deployment and monitoring of nodes on leading protocols, providing access to organizations looking to build their blockchain infrastructure with a high availability and low total cost of ownership.",
+  },
+  {
+    id: "0x99510bfc5d420cd3c41458f4692b1c3321908734",
+    name: "Stakely",
+    description:
+      "Stakely is a top Staking-as-a-Service provider specialized in Proof-of-Stake blockchains. With over 40 supported networks, we offer reliable and secure staking services for the best staking experience.\n",
+  },
+  {
+    id: "0x05e393ec11d0197a88d84457908edd4ff235f323",
+    name: "Stakin - 0% Fee",
+    description:
+      "Stakin (https://stakin.com) provides institutional-grade non-custodial staking services and Web3 infrastructure. We operate on more than 40 blockchain networks including Ethereum, with 10,000+ ETH validators.",
+  },
+  {
+    id: "0x481afcd21a3447a7a04c4afd968df877b76d2310",
+    name: "Sun and Moon Collection",
+    description: "Celestial gains for celestial bodies.",
+  },
+  {
+    id: "0x089a97a8bc0c0f016f89f9cf42181ff06afb2daf",
+    name: "validator service",
+    description: "reliable validation",
+  },
+  {
+    id: "0x96fb413349cb4ec17410df929898ae9c0e56ae40",
+    name: "Blockshard Pioneers Vault ",
+    description:
+      "Leading vault featuring a strategic blend of DVT and vanilla validators. Around 50% are managed with SSV and Obol, employing highly skilled and experienced operators. \n\nThe remaining 50% are run on our advanced, high-availability system. The validator client is connected to five Ethereum nodes and uses Web3Signer to protect validator keys, ensuring enhanced reliability and security.\n\nA competitive fee of only 3.5% on rewards makes this vault a prudent choice for stakers. ",
+  },
+  {
+    id: "0x1b3ce55dde0e0d4b9a200855406e7b14334c10b0",
+    name: "Allnodes",
+    description:
+      "Reliable non-custodial Validator run by the industry leader - Allnodes",
+  },
+  {
+    id: "0x91211a4965e75152cb549b308f8ba398c3ab337e",
+    name: "Verse ETH Staking",
+    description:
+      "Earn rewards by staking your ETH in the Verse Ecosystem vault",
+  },
+  {
+    id: "0xdbdee04c72a02a740b9f26ada9203582c8a99daf",
+    name: "Chorus One - Obol DV vault",
+    description:
+      'This vault represents a distributed validator cluster , using Obol\'s DV technology to run the validators across multiple nodes. As a staker, you can expect higher uptime, decreased slashing risk, and a meaningful contribution to the decentralization of the Ethereum network.\n\nBy staking with an Obol DV, you automatically participate in the Obol Contributions initiative, contributing 1% of staking rewards to the "1% for Decentralization" retroactive funding model. Your contributions will be tracked and recognized by Obol and can be viewed on our staking dApp, OPUS Pool.',
+  },
+  {
+    id: "0x7eed3ea8d83ba4ccc1b20674f46825ece2fce594",
+    name: "Fox & Yaga",
+    description:
+      "Fox & Yaga is a collaboration vault on StakeWise dedicated to securing Ethereum while supporting the growth of Yaga, a unique jewelry brand. This initiative combines the strength of decentralized finance with a passion for fostering local businesses, making it a beacon of community-driven progress in the crypto space.",
+  },
+  {
+    id: "0x2ca762ed936e48c23b8a32403dd754020b07e86f",
+    name: "VincentTaglia.eth",
+    description:
+      "Independent Ethereum validator utilizing Nethermind+Teku, uncensored MEV relays, owned hardware, and the smoothing pool. Support decentralization and spread risk by staking with me. Learn more: https://vincenttaglia.eth.limo",
+  },
+  {
+    id: "0x64f2907f92631619ed7ea510982835f9e1024767",
+    name: "AutoStake ⚡️ 1% fee forever",
+    description:
+      "Slash insurance backed by a SAFU fund.\nDecentralized infrastructure powered by Distributed Validator Technology. \nConnect with us at: https://linktr.ee/autostake",
+  },
+  {
+    id: "0x302be829c61c287787030888bbcf11115ecd5773",
+    name: "Swiss Liquid Staking",
+    description:
+      "Unlock the potential of liquid Ethereum staking with Brick Towers, where we bring Swiss precision and security to your Ethereum journey.  Our platform, originally designed for banks, is now open to everyone, offering a cost-effective and non-custodial way to access protocol rewards by securing the network.  At your choice, mint osETH and enter the world of DeFi (lending & borrowing, yield farming and restaking) to enhance your rewards. Join us and experience a hassle-free liquid staking solution. Learn more about us at 🔗 https://bricktowers.io ",
+  },
+  {
+    id: "0x9c29c571847a68a947acec8bacd303e36bc72ec5",
+    name: "01node StakeWise V3 Vault",
+    description:
+      "01node operarated Vault, 1% commision; read more on  https://staking.01node.com/",
+  },
+  {
+    id: "0x2a14b47e551288f3f5e35842970cc2bc78ceb47d",
+    name: "Cryptoknight Vault",
+    description:
+      "Validator nodes operated by the leading Web3 education community for working professionals in Singapore - Cryptoknight Academy. Limited to 100 validator keys, we work with the Stakesaurus home staking community for diversification of our physical infrastructure and clients. DVTs will be implemented for additional performance and security boosts as we scale up!",
+  },
+  {
+    id: "0xb0fd1783826ce1a254ba8d5c409b267238faa9cd",
+    name: "CommonWealth",
+    description:
+      "Professional setup powered by minority pairs and validator client strategies.\nCompetitive 3.9% operator fee.",
+  },
+  {
+    id: "0xe2d8f982708ce1e3814c8986cbab624ca926288a",
+    name: "KysenPool Zen 🧘",
+    description:
+      "Unlock the Power of Staking with Our Trusted Blockchain Infrastructure Services!  We are a Stakewise-verified validator.  Operating since 2019 as a qualified and elected validator in over a dozen mainnet chains, we have expanded our staking infrastructure into 3 continents.  This vault is managed by a multisig wallet and limited to 500 validators (16,000 ETH).  Visit our Website @ https://kysenpool.io -or- Web3 Guidebooks @ https://docs.kysenpool.io for more information.\n\nFor institutional ETH holders looking to have your own private Stakewise pool, you can establish a commercial agreement with us to have a custom vault fee, set up operational and reporting procedures, manage jurisdictional requirements, uphold security standards and maintain compliance with regulations.  Please reach out to our business development team at bizdev@kysenpool.io",
+  },
+  {
+    id: "0x649955f4189c3921df60e25f58cb1e81070fedb0",
+    name: "Stakesaurus SEA Home Nodes",
+    description:
+      "High availability setup using a network of home-based nodes in SEA running minority clients by Stakesaurus' community\n\nCentralised infrastructure is bad for both the network and your own ETH (e.g. cloud service, data centres). Correlation penalties can wipe out your hard-earned yield in the best case and take a sizeable chunk out of your capital in the worst.\n\nOur solution? A geographically distributed network of home-staking nodes in Singapore and Malaysia (other SEA locations coming soon!) contributed by the Stakesaurus community. This network of uncorrelated nodes act as fallbacks for one another to provide a high availability staking service!\n\nWe believe in keeping staking operations as simple as possible to minimise tail risk events and security flaws — e.g. Most slashing events are due to institutions over-engineering their setups.\n\nAt Stakesaurus, we are growing the solo-staking community in SEA with our open source content — https://linktr.ee/stakesaurus",
+  },
+  {
+    id: "0x06ec2d731b2fa8895e5a9482c22ee33573f44a13",
+    name: "Big Euler's e-Stake",
+    description:
+      "Inspired by Euler's number, we're all about exponential growth. Our fee is based off 'e' itself - just 2.71% - Solo Staker | Local Hardware | Lighthouse | Nethermind | Ultrasound Money | Flashbots | We also operate genesis Ethereum validator #4394 using the same system",
+  },
+  {
+    id: "0x30c13e93c170885e9969ae2e4ff5f020807e0d74",
+    name: "Simply Staking: Premium",
+    description:
+      "Minimum Fee - Unrestricted Maximum MEV - Zero-Cloud - Private Datacentre - 24/7 Engineering Support - All included in Simply Staking's Premium Vault...\n\n\n- Fee: Competitive 3.75%.\n- MEV: Extensive list of MEV relays, free of restrictions, with the ultimate goal of using a global set of relays to maximise profit.\n- Private Data Centre Operation: Your ETH is securely managed in our privately operated data center in Malta, prioritising security and reliability, free from the dependencies of cloud services.",
+  },
+  {
+    id: "0x5fe23da0835ecf01f4faf9b07fdcfd603dfbd146",
+    name: "Lavender.Five Nodes",
+    description:
+      "Fortifying crypto networks with MPC/SSV security, slash insurance, and open source contributions. Connect with us at https://linktr.ee/lavenderfive.",
+  },
+  {
+    id: "0x366e7f2e3462e31001560875b9877a0714e90d9c",
+    name: "A41",
+    description:
+      "A41 is a blockchain infrastructure company. We provide staking service on multichain. Stake your token with us now and earn rewards!",
+  },
+  {
+    id: "0xea6db44b0d7e20f72f2b4b3deaaa9e710e718bd9",
+    name: "DSRV",
+    description:
+      "DSRV is an integrated blockchain solutions company with the mission of enriching the crypto ecosystem via stronger connectivity. We strive to be your gateway to a suite of all-comprehensive blockchain services. Everything distributed, served complete.",
+  },
+  {
+    id: "0x907660805fb78ae4e8419f76537671cc4df2d4ab",
+    name: "Peer Ventures",
+    description:
+      "Peer Ventures Vault is managed by experienced and conscientious node operators with on-premises equipment. We strive for the top tier of effectiveness with reliable power and internet, battery backups, proper maintenance, and monitoring. We've been operating highly effective validators on mainnet since January 2022, and manage 5,000 Holesky validators from genesis, for the lifetime of the testnet.  Ethereum is all we do, and the health of the beacon chain is important to us. This vault is self-limited to 100,000 ETH, which is less than 0.5% of the stake at time of creation.",
+  },
+  {
+    id: "0x3102b4013cb506481e959c8f4500b994d2bff22e",
+    name: "BRAZILPRACIMA",
+    description: "Independent solo operator since genesis",
+  },
+  {
+    id: "0xa1e229db735f7aa3882bfd22fa6b89734225a3d1",
+    name: "Amamu.io",
+    description:
+      "Amamu is an independent, secure, and reliable staking provider founded in 2020 with the primary aim of accelerating decentralization throughout the blockchain landscape. ",
+  },
+  {
+    id: "0x41adbf980d09e6e50afffe606e5c259dcec0af27",
+    name: "Luganodes - Stakekit",
+    description: null,
+  },
+  {
+    id: "0xf124ef4e47eefa1d6b1cb5f835ef1669e9a29a7c",
+    name: "Orangefin Ventures",
+    description:
+      "Orangefin Ventures is one of the world's leading node operators with nearly $100M assets under delegation. Having almost a decade of experience in the cryptocurrency space, Orangefin Ventures operates some of the most technically sophisticated infrastructure that secures several blockchains and spans across public clouds and our own datacenters (including our own ASN and public IPv4 range). To learn more, visit https://www.orangefin.ventures!",
+  },
+] as const