+++ sdk/src/protocols/stakeWise/v3/_gnoVaults.ts
@@ -1,0 +1,120 @@
+// This file is auto-generated. Do not edit!
+
+export default [
+  {
+    id: "0x4b4406ed8659d03423490d8b62a1639206da0a7a",
+    name: "Genesis Vault",
+    description:
+      "This is the vault for the StakeWise V2 pool. After migrating from StakeWise V2 to V3, you will be staking in this vault. The node operators in this vault are Chorus One, Cryptomanufaktur, Finoa, Serenita, and StakeWise Labs.",
+  },
+  {
+    id: "0x2cd404d9d75436e7d6ddbccc2fb9cf7c06941bf1",
+    name: "NEDO",
+    description:
+      "Welcome to the Nedo Vault on StakeWise v3!\n\nWhy Stake with Nedo?\n\nMaximized Returns: Nedo is dedicated to boosting your APY through operational excellence and cost efficiency. We meticulously manage operational costs to enhance your returns and accelerate asset growth.\n\nUncompromised Security: Security is our top priority. We employ cutting-edge measures to safeguard your investments, making Nedo a trusted platform for both institutional and individual investors.\n\nUnwavering Stability: By reinvesting operational fees into our infrastructure, we ensure continuous uptime and reliable performance. Stake with confidence knowing your assets are in steady hands.\n\nTransparency: We believe in openness. Explore our infrastructure in real-time on our transparency dashboard at nedo.fm/gnosis\n\nJoin Nedo Today: Experience secure and profitable staking on the Gnosis Chain with StakeWise v3. Let your assets work harder for you with Nedo.\n\nContact us at: gm@nedo.fm",
+  },
+  {
+    id: "0x00025c729a3364faef02c7d1f577068d87e90ba6",
+    name: "Serenita",
+    description:
+      "Serenita provides enterprise-grade staking services, prioritizing optimal performance while mitigating the risks associated with using majority clients. Our exclusive focus is on the Ethereum ecosystem and we are dedicated to its long-term sustainability and neutrality.\n\nWe manage one of the best StakeWise Vaults on Ethereum, and aim to be a leading staking provider on Gnosis Chain as well.\n\nFor a deeper understanding of our operational approach, visit our website at serenita.io",
+  },
+  {
+    id: "0x33c346928ed9249cf1d5fc16ae32a8cffa1671ad",
+    name: "Axol.io",
+    description:
+      "Sustainable nodes for gnosis, ran by the LATAM team at Axol.io",
+  },
+  {
+    id: "0xe0a342ed4e0f0dbe97c4810534cfcb6550ea017d",
+    name: "Stakecat",
+    description:
+      "Node operators, securing and decentralizing Ethereum |  AVS operator: http://avs.stakecat.space\n \nWe provide greater decentralization for networks with our own infrastructure for validators and high availability to compete with centralized data center performance.\n\nmore about Stakecat here: https://www.stakecat.space/",
+  },
+  {
+    id: "0x809484b8579dc605917b8f94aa284282d5fe375d",
+    name: "IP Stake",
+    description:
+      "I'm a Gnosis Chain solo operator based in Europe since January 2022. I run on-premises high availability setup with 99.9% uptime. I'm a verified nodeset.io operator since June 2024, a verified member of stakersunion.com and a Lido SimpleDVT module mainet operator since July 2024. Stake with me!",
+  },
+  {
+    id: "0x8898a1199a36023e9791f445bbf498755a180b7f",
+    name: "H2O Nodes",
+    description:
+      "Help decentralize Ethereum by staking with us - we run minority clients (usually Nimbus + Nethermind) running on bare metal servers in minority data centers. \n\nRead more at www.h2o-nodes.com\n\nHappy staking! ",
+  },
+  {
+    id: "0x3cb4692177525db38d983da0445d4eb25c3826de",
+    name: "Stakesaurus SEA Home Nodes",
+    description:
+      "High availability setup using a network of home-based nodes in SEA running minority clients by Stakesaurus' community. Centralised infrastructure is bad for both the network and your own GNO (e.g. cloud service, data centres). Correlation penalties can wipe out your hard-earned yield in the best case and take a sizeable chunk out of your capital in the worst. Our solution? A geographically distributed network of home-staking nodes in Singapore and Malaysia (other SEA locations coming soon!) contributed by the Stakesaurus community. This network of uncorrelated nodes act as fallbacks for one another to provide a high availability staking service! At Stakesaurus, we are growing the solo-staking community in SEA with our open source content — https://linktr.ee/stakesaurus",
+  },
+  {
+    id: "0x27fbf9cc8a4dbf438aecb32f278e287f157a267b",
+    name: "alpineforge.eu gnosis vault",
+    description:
+      "stakewise V3 vault on gnosis, brought to you by alpineforge.eu with <3\nhappy staking! 🦉hoot, hoot!",
+  },
+  {
+    id: "0xe32773a0f6fa8ebf2a32c040dacd031a1ad1e530",
+    name: "Orangefin Ventures",
+    description:
+      "Orangefin Ventures is one of the world's leading node operators with nearly $100M assets under delegation. Having almost a decade of experience in the cryptocurrency space, Orangefin Ventures operates some of the most technically sophisticated infrastructure that secures several blockchains and spans across public clouds and our own datacenters (including our own ASN and public IPv4 range). To learn more, visit https://www.orangefin.ventures!",
+  },
+  {
+    id: "0x0f6f727421ed6229fcac821147ae104c9326229e",
+    name: "Dappnode vault",
+    description:
+      "This is a vault from a home node runner using a Dappnode secured by an UPS and connected via FTTH. Everybody is welcome. 5% Fees",
+  },
+  {
+    id: "0x27b7cb7bc376c0798651238ce5294f70b283bc61",
+    name: "Blockscape Gnosis Vault ",
+    description:
+      "Institutional-grade staking for everyone made by Blockscape in Switzerland. As staker, you benefit from a low commission fee of just 12.5%, while still enjoying the same levels of security and reliability. Our unique setup provides highest standards in security and reliability. With more than 10.000 stakers and over $500.000.000 under management, Blockscape is one of the most reliable staking providers in the world. Moreover, we offer dedicated vaults for institutions with specific requirements. Reach out to us at: partner@blockscape.network or visit https://blockscape.network for more information.",
+  },
+  {
+    id: "0x8a20669daa7969788f0ae7dbd38ac76e102ef213",
+    name: "FCC GNO Vault",
+    description: "FCC GNO Liquid Staking Public Vault",
+  },
+  {
+    id: "0xc554bdbbed28d34e3675c6c53dc0ab645cf285e0",
+    name: "Wunode",
+    description: "Dappnode vault, fiber gigabit uplink, UPS",
+  },
+  {
+    id: "0x9cb24ec88fca91aff9def12989ab9274f3992d3e",
+    name: "Vespertine",
+    description: "Investing in a bankless future.",
+  },
+  {
+    id: "0x5fa6aa1ebbbbcbfd8757b6f56abbe468ac3a6026",
+    name: "Stakers Union Vault",
+    description:
+      "The home-stakers collective:\nEmpowering individual stakers, preserving decentralization across Ethereum-aligned networks.",
+  },
+  {
+    id: "0x20dd41e898a58ee4148a2f6db2c2be7082441932",
+    name: "Sophia's Gnosis",
+    description:
+      "Whispers of Sophia's Gnosis\n\nIn realms beyond the veil of sight,\nI dance with truths both dark and light.\nSophia's wisdom, my guiding star,\nGnosis beckons from afar.\n\nThrough winding paths of mystic lore,\nI seek the keys to cosmic doors.\nEach step unveils a sacred sign,\nAs mortal flesh and soul entwine.\n\nIn silence deep, I hear the call,\nOf hidden knowledge, one and all.\nThe serpent's whisper, soft and low,\nReveals the truths I've come to know.\n\nIllumination floods my mind,\nAs earthly shackles fall behind.\nIn gnosis found, both pure and bright,\nI rise, transformed, into the light.\n\nFor wisdom's flame burns ever true,\nIn hearts that dare to see anew.\nAnd in this knowing, vast and free,\nI find the cosmic self that sleeps in me.",
+  },
+  {
+    id: "0x0ca22f2d5ad17142bb1065ae828d9db28f86ca1c",
+    name: "Thot",
+    description: "Testing vaults ",
+  },
+  {
+    id: "0x2435992a2bf5feaa535d16edaad3d525dc94e0e8",
+    name: "Forsaken",
+    description:
+      "Home Dappnode staker based in Greece since 2022.\nProviding your worst possible vault choice. If there is a power outage then we are both in pain.\nEnjoy!",
+  },
+  {
+    id: "0xdb57ca4f12ad6b3f962103ba5c00b6d967b9b8ac",
+    name: "NotMyVault",
+    description:
+      "Just exploring the possibilities !\n\nPLease don't put your GNO in here. Just testing!",
+  },
+] as const